Release v6.4 - Dragonflight Season 2 dungeons

+++ app/Http/Controllers/APIPridefulEnemyController.php
@@ -8,6 +8,7 @@ use App\Models\DungeonRoute;
 use App\Models\Enemies\PridefulEnemy;
 use App\Models\Enemy;
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\Routing\ResponseFactory;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
@@ -18,25 +19,23 @@ class APIPridefulEnemyController extends Controller
 {
     /**
      * @param Request $request
-     * @param DungeonRoute $dungeonroute
+     * @param DungeonRoute $dungeonRoute
      * @param Enemy $enemy
      * @return PridefulEnemy
      * @throws Exception
      */
-    function store(Request $request, DungeonRoute $dungeonroute, Enemy $enemy)
+    function store(Request $request, DungeonRoute $dungeonRoute, Enemy $enemy)
     {
-        if (!$dungeonroute->isSandbox()) {
-            $this->authorize('edit', $dungeonroute);
-        }
+        $this->authorize('edit', $dungeonRoute);
 
         /** @var PridefulEnemy $pridefulEnemy */
-        $pridefulEnemy = PridefulEnemy::where('dungeon_route_id', $dungeonroute->id)->where('enemy_id', $enemy->id)->first();
+        $pridefulEnemy = PridefulEnemy::where('dungeon_route_id', $dungeonRoute->id)->where('enemy_id', $enemy->id)->first();
 
         if ($pridefulEnemy === null) {
             $pridefulEnemy = new PridefulEnemy();
         }
 
-        $pridefulEnemy->dungeon_route_id = $dungeonroute->id;
+        $pridefulEnemy->dungeon_route_id = $dungeonRoute->id;
         $pridefulEnemy->enemy_id         = (int)$enemy->id;
         $pridefulEnemy->floor_id         = (int)$request->get('floor_id');
         $pridefulEnemy->lat              = (float)$request->get('lat');
@@ -47,30 +46,33 @@ class APIPridefulEnemyController extends Controller
         }
 
         if (Auth::check()) {
-            broadcast(new ModelChangedEvent($dungeonroute, Auth::getUser(), $pridefulEnemy));
+            broadcast(new ModelChangedEvent($dungeonRoute, Auth::getUser(), $pridefulEnemy));
         }
 
-        $dungeonroute->touch();
+        $dungeonRoute->touch();
 
         return $pridefulEnemy;
     }
 
     /**
      * @param Request $request
-     * @param DungeonRoute $dungeonroute
+     * @param DungeonRoute $dungeonRoute
      * @param Enemy $enemy
      * @return Response|ResponseFactory
+     * @throws AuthorizationException
      */
-    function delete(Request $request, DungeonRoute $dungeonroute, Enemy $enemy)
+    function delete(Request $request, DungeonRoute $dungeonRoute, Enemy $enemy)
     {
+        $this->authorize('edit', $dungeonRoute);
+
         try {
             /** @var PridefulEnemy $pridefulEnemy */
-            $pridefulEnemy = PridefulEnemy::where('dungeon_route_id', $dungeonroute->id)->where('enemy_id', $enemy->id)->first();
+            $pridefulEnemy = PridefulEnemy::where('dungeon_route_id', $dungeonRoute->id)->where('enemy_id', $enemy->id)->first();
             if ($pridefulEnemy && $pridefulEnemy->delete() && Auth::check()) {
-                broadcast(new ModelDeletedEvent($dungeonroute, Auth::getUser(), $pridefulEnemy));
+                broadcast(new ModelDeletedEvent($dungeonRoute, Auth::getUser(), $pridefulEnemy));
             }
 
-            $dungeonroute->touch();
+            $dungeonRoute->touch();
 
             $result = response()->noContent();
         } catch (Exception $ex) {
